using Firestore ın contact app
This project is the same as the one I previously created, which was done using Sqlite (Room). However, for this project, I utilized a Firebase (Firestore ). Instead of sharing the entire project folders that are identical, I'm sharing only the files where modifications were made (specifically, the datasource package, entity package, di package, ).